+++ PhotosAlbum/PhotosAlbum/Source/View/PhotoCollectionDelegateFlowLayout.swift
@@ -1,0 +1,34 @@
+//
+//  PhotoCollectionDelegate.swift
+//  PhotosAlbum
+//
+//  Created by YEONGJIN JANG on 2022/08/26.
+//
+
+import UIKit
+
+private enum LayoutConstant {
+    static let spacing: CGFloat = 2
+    static let itemHeight: CGFloat = 80
+}
+
+final class PhotoCollectionDelegateFlowLayout: NSObject, UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        print(indexPath.row)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        let width = itemWidth(for: UIScreen.main.bounds.width, spacing: LayoutConstant.spacing)
+        return CGSize(width: width, height: LayoutConstant.itemHeight)
+    }
+    
+    private func itemWidth(for width: CGFloat, spacing: CGFloat) -> CGFloat {
+        let itemsInRow: CGFloat = 4
+        let totalSpacing:  CGFloat = 2 * spacing + (itemsInRow - 1) * spacing
+        let finalWidth = (width - totalSpacing) / itemsInRow
+        
+        return finalWidth - 5.0
+    }
+    
+    
+}
